+++ lib/features/home/presentation/bloc/home_state.dart
@@ -6,6 +6,10 @@ enum HomeBlocStatus {
   loaded,
   error,
   repositorySelected,
+  askForCloningRepository,
+  cloneSuccess,
+  cloning,
+  cloneProgress,
 }
 
 @MappableClass()
@@ -13,10 +17,18 @@ class HomeState with HomeStateMappable {
   final HomeBlocStatus status;
   final String currentRepositoryName;
   final String repositoryPath;
+  final String errorMessage;
+  final String cloneDestinationPath;
+  final String cloneRepositoryUrl;
+  final double cloneProgress;
 
   const HomeState({
     this.status = HomeBlocStatus.initial,
     this.currentRepositoryName = "",
     this.repositoryPath = "",
+    this.errorMessage = "",
+    this.cloneDestinationPath = "",
+    this.cloneRepositoryUrl = "",
+    this.cloneProgress = 0.0,
   });
 }